Add a defaultOptions() method
I have run into a scenario where I need to add additional `curl_setopt()` configurations in order for my environment to properly communicate with the VoiceIt API. One workaround that I have made locally is moving most of the repeating `curl_setopt()` calls to a `setDefaultOptions()` function which is then overrideable to have additional options be used. In my case, it was the need to disable SSL checks for one of my dev instances so that it would proceed down the network pipeline to the VoiceIt API.

```
      curl_setopt($crl, CURLOPT_SSL_VERIFYHOST, false);
      curl_setopt($crl, CURLOPT_SSL_VERIFYPEER, false);
```

I'd prefer to not have to manage my own a locally-modified version of this class for the long-term. Can this type of behavior I've described and submitted be added in to the main branch?